Add a new column to keep track of tests
This adds a new column to show if the current proposal already have tests and provide a link to the respective issue/PR on Test262.

It should be checked the tests are already merged at completion.

For other items on stage 2, I'm adding a link to create a new issue, but I can remove that and live a blank space if it improves reading.